requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.2.1
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.13.1
shapely==2.0.6
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
